backend/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
decorator==5.1.1
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
fonttools==4.31.2
idna==3.3
joblib==1.1.0
kiwisolver==1.4.1
matplotlib==3.5.1
mlxtend==0.19.0
numpy==1.22.3
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
PyJWT==2.3.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sqlparse==0.4.2
threadpoolctl==3.1.0
tzdata==2021.5
urllib3==1.26.9
